// delivery_toop-admin/frontend/src/models/finance/subordinate.ts
import { Person } from './../person';
import { Company } from './../company/company';
import { City } from './../city';
import { State } from './../state';
export class Subordinate {
  id?: number;
  _id?: string;
  uf: State;
  city: City;
  legalPerson: Company;
  physicalPerson: Person;
  socialReason: string;
  cnpj: number;
  fantasyName: string;
  representativesName: string;
  representativeEmail: string;
  representativePhone: number;
  name: string;
  cpf: number;
  email: string;
  phone: number;
  dateOfBirth: string;
  cep: number;
  publicPlace: string;
  number: number;
  complement: string;
  neighborhood: string;
  accountType: string;
  bank: string;
  agency: number;
  agencyDigit: number;
  account: number;
  accountDigit: number;
  fees: string;
  tariff: number;
  percentage: number;
  legal: any;
  physical: any;
  address: any;
  feesTypes: any;
}